requirements.txt
flask==2.3.2
openai==0.27.2
python-dotenv==1.0.0
